+++ src/abstractarray.jl
@@ -1,31 +1,22 @@
-abstract type AbstractAccArray{T, N} <: DenseArray{T, N} end
+# Dense GPU Array
+abstract type GPUArray{T, N} <: DenseArray{T, N} end
+
 # Sampler type that acts like a texture/image and allows interpolated access
-abstract type AbstractSampler{T, N} <: AbstractAccArray{T, N} end
+abstract type Sampler{T, N} <: DenseArray{T, N} end
 
-const AccVector{T} = AbstractAccArray{T, 1}
-const AccMatrix{T} = AbstractAccArray{T, 2}
-const AccVecOrMat{T} = Union{AbstractAccArray{T, 1}, AbstractAccArray{T, 2}}
-
-
-
-type GPUArray{T, N, B, C} <: AbstractAccArray{T, N}
-    buffer::B
-    size::NTuple{N, Int}
-    context::C
-end
-
-
-# interfaces
+const GPUVector{T} = GPUArray{T, 1}
+const GPUMatrix{T} = GPUArray{T, 2}
+const GPUVecOrMat{T} = Union{GPUArray{T, 1}, GPUArray{T, 2}}
 
 #=
 Interface for accessing the lower level
 =#
-buffer(A::AbstractAccArray) = A.buffer
-context(A::AbstractAccArray) = A.context
+buffer(A::GPUArray) = A.buffer
+context(A::GPUArray) = A.context
 default_buffer_type(typ, context) = error("Found unsupported context: $context")
 
 # GPU Local Memory
-immutable LocalMemory{T} <: AbstractAccArray{T, 1}
+immutable LocalMemory{T} <: GPUArray{T, 1}
     size::Int
 end
 
@@ -36,31 +27,29 @@ linear index in a GPU kernel
 function linear_index end
 
 
-
 #=
 AbstractArray interface
 =#
-Base.eltype{T}(::AbstractAccArray{T}) = T
-Base.size(A::AbstractAccArray) = A.size
+Base.eltype{T}(::GPUArray{T}) = T
+Base.size(A::GPUArray) = A.size
 
-function Base.show(io::IO, mt::MIME"text/plain", A::AbstractAccArray)
-    println(io, "GPUArray with ctx: $(context(A)): ")
+function Base.show(io::IO, mt::MIME"text/plain", A::GPUArray)
     show(io, mt, Array(A))
 end
-function Base.showcompact(io::IO, mt::MIME"text/plain", A::AbstractAccArray)
+function Base.showcompact(io::IO, mt::MIME"text/plain", A::GPUArray)
     showcompact(io, mt, Array(A))
 end
 
-function Base.similar{T <: AbstractAccArray}(x::T)
+function Base.similar{T <: GPUArray}(x::T)
     similar(x, eltype(x), size(x))
 end
-function Base.similar{T <: AbstractAccArray, ET}(x::T, ::Type{ET}; kw_args...)
+function Base.similar{T <: GPUArray, ET}(x::T, ::Type{ET}; kw_args...)
     similar(x, ET, size(x); kw_args...)
 end
-function Base.similar{T <: AbstractAccArray, N}(x::T, dims::NTuple{N, Int}; kw_args...)
+function Base.similar{T <: GPUArray, N}(x::T, dims::NTuple{N, Int}; kw_args...)
     similar(x, eltype(x), dims; kw_args...)
 end
-function Base.similar{N, ET}(x::AbstractAccArray, ::Type{ET}, sz::NTuple{N, Int}; kw_args...)
+function Base.similar{N, ET}(x::GPUArray, ::Type{ET}, sz::NTuple{N, Int}; kw_args...)
     similar(typeof(x), ET, sz, context = context(x); kw_args...)
 end
 
@@ -78,109 +67,36 @@ end
 #=
 Host to Device data transfers
 =#
-function (::Type{A}){A <: AbstractAccArray}(x::AbstractArray)
+function (::Type{A}){A <: GPUArray}(x::AbstractArray)
     A(collect(x))
 end
-function (::Type{A}){A <: AbstractAccArray}(x::Array; kw_args...)
+function (::Type{A}){A <: GPUArray}(x::Array; kw_args...)
     out = similar(A, eltype(x), size(x); kw_args...)
     copy!(out, x)
     out
 end
-Base.convert{A <: AbstractAccArray}(::Type{A}, x::AbstractArray) = A(x)
-Base.convert{A <: AbstractAccArray}(::Type{A}, x::A) = x
+Base.convert{A <: GPUArray}(::Type{A}, x::AbstractArray) = A(x)
+Base.convert{A <: GPUArray}(::Type{A}, x::A) = x
 
 #=
 Device to host data transfers
 =#
-function (::Type{Array}){T, N}(device_array::AbstractAccArray{T, N})
+function (::Type{Array}){T, N}(device_array::GPUArray{T, N})
     Array{T, N}(device_array)
 end
-function (AT::Type{Array{T, N}}){T, N}(device_array::AbstractAccArray)
+function (AT::Type{Array{T, N}}){T, N}(device_array::GPUArray)
     convert(AT, Array(device_array))
 end
-function (AT::Type{Array{T, N}}){T, N}(device_array::AbstractAccArray{T, N})
+function (AT::Type{Array{T, N}}){T, N}(device_array::GPUArray{T, N})
     hostarray = similar(AT, size(device_array))
     copy!(hostarray, device_array)
     hostarray
 end
 
 
-# Function needed to be overloaded by backends
-function mapidx end
-# same for mapreduce
-function acc_mapreduce end
-
-
 ######################################
 # Broadcast
 include("broadcast.jl")
-
-
-# we need to overload all the different broadcast functions, since x... is ambigious
-
-# TODO check size
-function Base.map!(f::Function, A::AbstractAccArray, args::AbstractAccArray...)
-    broadcast!(f, A, args...)
-end
-function Base.map(f::Function, A::AbstractAccArray, args::AbstractAccArray...)
-    broadcast(f, A, args...)
-end
-
-
-#############################
-# reduce
-
-# hack to get around of fetching the first element of the GPUArray
-# as a startvalue, which is a bit complicated with the current reduce implementation
-function startvalue(f, T)
-    error("Please supply a starting value for mapreduce. E.g: mapreduce($f, $op, 1, A)")
-end
-startvalue(::typeof(+), T) = zero(T)
-startvalue(::typeof(*), T) = one(T)
-startvalue(::typeof(Base.scalarmin), T) = typemax(T)
-startvalue(::typeof(Base.scalarmax), T) = typemin(T)
-
-# TODO widen and support Int64 and use Base.r_promote_type
-gpu_promote_type{T}(op, ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(op, ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(+), ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(*), ::Type{T}) = T
-gpu_promote_type{T<:Number}(::typeof(+), ::Type{T}) = typeof(zero(T)+zero(T))
-gpu_promote_type{T<:Number}(::typeof(*), ::Type{T}) = typeof(one(T)*one(T))
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(Base.scalarmax), ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(Base.scalarmin), ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(max), ::Type{T}) = T
-gpu_promote_type{T<:Base.WidenReduceResult}(::typeof(min), ::Type{T}) = T
-
-function Base.mapreduce{T, N}(f::Function, op::Function, A::AbstractAccArray{T, N})
-    OT = gpu_promote_type(op, T)
-    v0 = startvalue(op, OT) # TODO do this better
-    mapreduce(f, op, v0, A)
-end
-
-
-function Base.mapreduce(f, op, v0, A::AbstractAccArray, B::AbstractAccArray, C::Number)
-    acc_mapreduce(f, op, v0, A, (B, C))
-end
-function Base.mapreduce(f, op, v0, A::AbstractAccArray, B::AbstractAccArray)
-    acc_mapreduce(f, op, v0, A, (B,))
-end
-function Base.mapreduce(f, op, v0, A::AbstractAccArray)
-    acc_mapreduce(f, op, v0, A, ())
-end
-
-############################################
-# Constructor
-
-function Base.fill!{T, N}(A::AbstractAccArray{T, N}, val)
-    valconv = T(val)
-    gpu_call(const_kernel2, A, (A, valconv, Cuint(length(A))))
-    A
-end
-function Base.rand{T <: AbstractAccArray, ET}(::Type{T}, ::Type{ET}, size...)
-    T(rand(ET, size...))
-end
-
 
 ############################################
 # serialization
@@ -202,11 +118,9 @@ function Base.deserialize{T <: GPUArray}(s::BaseSerializer, ::Type{T})
     T(A)
 end
 
-import Base: copy!, getindex, setindex!
-
 @inline unpack_buffer(x) = x
-@inline unpack_buffer(x::AbstractAccArray) = buffer(x)
-@inline unpack_buffer(x::Ref{<: AbstractAccArray}) = unpack_buffer(x[])
+@inline unpack_buffer(x::GPUArray) = buffer(x)
+@inline unpack_buffer(x::Ref{<: GPUArray}) = unpack_buffer(x[])
 
 function to_cartesian(A, indices::Tuple)
     start = CartesianIndex(ntuple(length(indices)) do i
@@ -245,7 +159,7 @@ function array_convert{T, N, T2}(t::Type{Array{T, N}}, x::T2)
     return reshape(map(T, arr), dims) # broadcast dims
 end
 
-for (D, S) in ((AbstractAccArray, AbstractArray), (AbstractArray, AbstractAccArray), (AbstractAccArray, AbstractAccArray))
+for (D, S) in ((GPUArray, AbstractArray), (AbstractArray, GPUArray), (GPUArray, GPUArray))
     @eval begin
         function copy!(
                 dest::$D, doffset::Integer,
@@ -310,8 +224,8 @@ function copy_kernel!(state, dest, dest_offsets, src, src_offsets, shape, shape_
 end
 
 function copy!{T, N}(
-        dest::AbstractAccArray{T, N}, destcrange::CartesianRange{CartesianIndex{N}},
-        src::AbstractAccArray{T, N}, srccrange::CartesianRange{CartesianIndex{N}}
+        dest::GPUArray{T, N}, destcrange::CartesianRange{CartesianIndex{N}},
+        src::GPUArray{T, N}, srccrange::CartesianRange{CartesianIndex{N}}
     )
     shape = size(destcrange)
     if shape != size(srccrange)
@@ -331,7 +245,7 @@ end
 
 
 function copy!{T, N}(
-        dest::AbstractAccArray{T, N}, destcrange::CartesianRange{CartesianIndex{N}},
+        dest::GPUArray{T, N}, destcrange::CartesianRange{CartesianIndex{N}},
         src::AbstractArray{T, N}, srccrange::CartesianRange{CartesianIndex{N}}
     )
     # Is this efficient? Maybe!
@@ -347,7 +261,7 @@ end
 
 function copy!{T, N}(
         dest::AbstractArray{T, N}, destcrange::CartesianRange{CartesianIndex{N}},
-        src::AbstractAccArray{T, N}, srccrange::CartesianRange{CartesianIndex{N}}
+        src::GPUArray{T, N}, srccrange::CartesianRange{CartesianIndex{N}}
     )
     # Is this efficient? Maybe!
     dest_gpu = similar(src, size(destcrange))
@@ -357,102 +271,7 @@ function copy!{T, N}(
     dest
 end
 
-
-Base.copy(x::AbstractAccArray) = identity.(x)
-
-indexlength(A, i, array::AbstractArray) = length(array)
-indexlength(A, i, array::Number) = 1
-indexlength(A, i, array::Colon) = size(A, i)
-
-function Base.setindex!{T, N}(A::AbstractAccArray{T, N}, value, indexes...)
-    # similarly, value should always be a julia array
-    shape = ntuple(Val{N}) do i
-        indexlength(A, i, indexes[i])
-    end
-    if !isa(value, T) # TODO, shape check errors for x[1:3] = 1
-        Base.setindex_shape_check(value, indexes...)
-    end
-    checkbounds(A, indexes...)
-    v = array_convert(Array{T, N}, value)
-    # since you shouldn't update GPUArrays with single indices, we simplify the interface
-    # by always mapping to ranges
-    ranges_dest = to_cartesian(A, indexes)
-    ranges_src = CartesianRange(size(v))
-
-    copy!(A, ranges_dest, v, ranges_src)
-    return
-end
-
-
-
-
-
-function Base.getindex{T}(A::AbstractAccArray{T, 0})
-    Array(A)[]
-end
-
-function Base.getindex{T, N}(A::AbstractAccArray{T, N}, indexes...)
-    cindexes = Base.to_indices(A, indexes)
-    # similarly, value should always be a julia array
-    # We shouldn't really bother about checkbounds performance, since setindex/getindex will always be relatively slow
-    checkbounds(A, cindexes...)
-
-    shape = map(length, cindexes)
-    result = Array{T, length(shape)}(shape)
-    ranges_src = to_cartesian(A, cindexes)
-    ranges_dest = CartesianRange(shape)
-    copy!(result, ranges_dest, A, ranges_src)
-    if all(i-> isa(i, Integer), cindexes) # scalar
-        return result[]
-    else
-        return result
-    end
-end
-
-
-#Broadcast
-Base.@propagate_inbounds broadcast_index(::Val{false}, arg, shape, i) = arg
-Base.@propagate_inbounds function broadcast_index{T, N}(
-        ::Val{true}, arg::AbstractArray{T, N}, shape::NTuple{N, Integer}, i
-    )
-    @inbounds return arg[i]
-end
-@generated function broadcast_index{T, N}(::Val{true}, arg::AbstractArray{T, N}, shape, i)
-    idx = []
-    for i = 1:N
-        push!(idx, :(s[$i] < shape[$i] ? 1 : idx[$i]))
-    end
-    expr = quote
-        $(Expr(:meta, :inline, :propagate_inbounds))
-        s = size(arg)
-        idx = ind2sub(shape, i)
-        @inbounds return arg[$(idx...)]
-    end
-end
-Base.@propagate_inbounds broadcast_index(arg, shape, i) = arg
-Base.@propagate_inbounds function broadcast_index{T, N}(
-        arg::AbstractArray{T, N}, shape::NTuple{N, Integer}, i
-    )
-    return arg[i]
-end
-Base.@propagate_inbounds function broadcast_index(
-        arg::AbstractArray, shape::Integer, i::Integer
-    )
-    return arg[i]
-end
-@generated function broadcast_index{T, N}(arg::AbstractArray{T, N}, shape, i)
-    idx = []
-    for i = 1:N
-        push!(idx, :(ifelse(s[$i] < shape[$i], 1, idx[$i])))
-    end
-    expr = quote
-        $(Expr(:meta, :inline, :propagate_inbounds))
-        s = size(arg)
-        idx = ind2sub(shape, i)
-        @inbounds return arg[$(idx...)]
-    end
-end
-
+Base.copy(x::GPUArray) = identity.(x)
 
 #=
 reinterpret taken from julia base/array.jl
@@ -474,20 +293,20 @@ This makes it easier to do checks just on the high level.
 """
 function unsafe_reinterpret end
 
-function reinterpret(::Type{T}, a::AbstractAccArray{S,1}) where T where S
+function reinterpret(::Type{T}, a::GPUArray{S,1}) where T where S
     nel = Int(div(length(a)*sizeof(S),sizeof(T)))
     # TODO: maybe check that remainder is zero?
     return reinterpret(T, a, (nel,))
 end
 
-function reinterpret(::Type{T}, a::AbstractAccArray{S}) where T where S
+function reinterpret(::Type{T}, a::GPUArray{S}) where T where S
     if sizeof(S) != sizeof(T)
         throw(ArgumentError("result shape not specified"))
     end
     reinterpret(T, a, size(a))
 end
 
-function reinterpret(::Type{T}, a::AbstractAccArray{S}, dims::NTuple{N,Int}) where T where S where N
+function reinterpret(::Type{T}, a::GPUArray{S}, dims::NTuple{N,Int}) where T where S where N
     if !isbits(T)
         throw(ArgumentError("cannot reinterpret Array{$(S)} to ::Type{Array{$(T)}}, type $(T) is not a bits type"))
     end
@@ -501,36 +320,9 @@ function reinterpret(::Type{T}, a::AbstractAccArray{S}, dims::NTuple{N,Int}) whe
     unsafe_reinterpret(T, a, dims)
 end
 
-function Base.reshape(a::AbstractAccArray{T}, dims::NTuple{N,Int}) where T where N
+function Base.reshape(a::GPUArray{T}, dims::NTuple{N,Int}) where T where N
     if prod(dims) != length(a)
         throw(DimensionMismatch("new dimensions $(dims) must be consistent with array size $(length(a))"))
     end
     unsafe_reinterpret(T, a, dims)
-end
-
-
-
-function mapreducedim_kernel(state, f, op, R::AbstractArray{T1, N}, A::AbstractArray{T, N}, slice_size, sizeA, dim) where {T1, T, N}
-    ilin = Cuint(linear_index(R, state))
-    accum = zero(T1)
-    @inbounds for i = Cuint(1):slice_size
-        idx = N == dim ? (ilin, i) : (i, ilin)
-        i2d = gpu_sub2ind(sizeA, idx)
-        accum = op(accum, f(A[i2d]))
-    end
-    R[ilin] = accum
-    return
-end
-function Base._mapreducedim!(f, op, R::AbstractAccArray, A::AbstractAccArray)
-    sizeR = size(R)
-    if all(x-> x == 1, sizeR)
-        x = mapreduce(f, op, A)
-        copy!(R, reshape([x], sizeR))
-        return R
-    end
-    @assert count(x-> x == 1, sizeR) == (ndims(R) - 1) "Not implemented"
-    dim = findfirst(x-> x == 1, sizeR)
-    slice_size = size(A, dim)
-    gpu_call(mapreducedim_kernel, R, (f, op, R, A, Cuint(slice_size), Cuint.(size(A)), Cuint(dim)))
-    return R
 end
