--- test/opengl.jl
@@ -1,32 +1,0 @@
-using GPUArrays
-using Base.Test
-using GPUArrays: free
-using GLAbstraction
-
-ctx = GLBackend.init()
-
-
-# more complex function for broadcast
-function test{T}(a::T, b)
-    x = sqrt(sin(a) * b) / T(10.0)
-    y = T(33.0)x + cos(b)
-    y * T(10.0)
-end
-
-@testset "broadcast Float32" begin
-    A = GPUArray(rand(Float32, 40))
-
-    A .= identity.(10f0)
-    @test all(x-> x == 10f0, Array(A))
-
-    A .= identity.(0.5f0)
-    B = test.(A, 10f0)
-    @test all(x-> x ≈ test(0.5f0, 10f0), Array(B))
-    A .= identity.(2f0)
-    C = (*).(A, 10f0)
-    @test all(x-> x == 20f0, Array(C))
-    D = (*).(A, B)
-    @test all(x-> x ≈ test(0.5f0, 10f0) * 2, Array(D))
-    D .= (+).((*).(A, B), 10f0)
-    @test all(x-> x ≈ test(0.5f0, 10f0) * 2 + 10f0, Array(D))
-end
